@@ -1,10 +1,9 @@
 module github.com/cybozu-go/neco-gcp
 
-go 1.13
+go 1.16
 
 require (
 	cloud.google.com/go v0.71.0
-	cloud.google.com/go/logging v1.1.2 // indirect
 	cloud.google.com/go/pubsub v1.6.1
 	github.com/cybozu-go/log v1.6.0
 	github.com/cybozu-go/well v1.10.0
@@ -15,15 +14,16 @@ require (
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/pelletier/go-toml v1.3.0 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
-	github.com/rakyll/statik v0.1.7
 	github.com/slack-go/slack v0.6.5
 	github.com/spf13/afero v1.2.2 // indirect
 	github.com/spf13/cobra v1.0.0
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
 	github.com/spf13/viper v1.7.1
 	github.com/stretchr/testify v1.6.1 // indirect
+	golang.org/x/tools v0.0.0-20201105220310-78b158585360 // indirect
 	google.golang.org/api v0.35.0
 	google.golang.org/genproto v0.0.0-20201106154455-f9bfe239b0ba
+	google.golang.org/grpc v1.33.2 // indirect
 	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
 	gopkg.in/yaml.v2 v2.3.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200605160147-a5ece683394c // indirect
