@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"context"
-	"io/ioutil"
 	"os"
 
 	"github.com/cybozu-go/log"
@@ -21,7 +20,7 @@ var necotestCreateImageCmd = &cobra.Command{
 		necotestCfg := gcp.NecoTestConfig(projectID, zone)
 		cc := gcp.NewComputeCLIClient(necotestCfg, "vmx-enabled")
 		well.Go(func(ctx context.Context) error {
-			f, err := ioutil.TempFile("", "*.yml")
+			f, err := os.CreateTemp("", "*.yml")
 			if err != nil {
 				return err
 			}
