@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"regexp"
 	"syscall"
@@ -55,7 +54,7 @@ func enableXForwarding() error {
 	}
 	defer f.Close()
 
-	data, err := ioutil.ReadAll(f)
+	data, err := io.ReadAll(f)
 	if err != nil {
 		return err
 	}
@@ -65,7 +64,7 @@ func enableXForwarding() error {
 	if err != nil {
 		return err
 	}
-	return ioutil.WriteFile(destFile, replaced, st.Mode())
+	return os.WriteFile(destFile, replaced, st.Mode())
 }
 
 func mountHomeDisk(ctx context.Context) error {
@@ -75,7 +74,7 @@ func mountHomeDisk(ctx context.Context) error {
 	}
 	defer f.Close()
 
-	data, err := ioutil.ReadAll(f)
+	data, err := io.ReadAll(f)
 	if err != nil {
 		return err
 	}
